requirements.txt
asgiref==3.5.2
cloudinary==1.30.0
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==3.2.16
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests-oauthlib==1.3.1
sqlparse==0.4.3
